Add unit tests, Conversion in AggregateEntityConfiguration and minor changes.

+++ src/modules/budget/domain/BudgetLimit.cs
@@ -5,4 +5,4 @@
 /// </summary>
 /// <param name="Value">Value.</param>
 /// <param name="Currency">Currency.</param>
-public record class BudgetLimit(float Value, Currency Currency);
+public record class BudgetLimit(decimal Value, Currency Currency);
